restructuring (#9)
* restructuring

* wip
