+++ src/App.js
@@ -5,17 +5,14 @@ import BurgerBuilder from './containers/BurgerBuilder/BurgerBuilder'
 import Checkout from './containers/Checkout/Checkout'
 import Orders from './containers/Orders/Orders'
 
-
-const App = () => {
-  return (
-    <Layout>
-      <Switch>
-        <Route path='/checkout' component={Checkout} />
-        <Route path='/orders' component={Orders} />
-        <Route path='/' exact component={BurgerBuilder} />
-      </Switch>
-    </Layout>
-  )
-}
+const App = () => (
+  <Layout>
+    <Switch>
+      <Route path='/checkout' component={Checkout} />
+      <Route path='/orders' component={Orders} />
+      <Route path='/' exact component={BurgerBuilder} />
+    </Switch>
+  </Layout>
+)
  
 export default App
