+++ src/assets/images/index.js
@@ -1,15 +1,15 @@
 export const images = {
   ingredients: {
-    bacon: require('./gallery/bacon.svg'),
-    cheese: require('./gallery/cheese.svg'),
-    meat: require('./gallery/meat.svg'),
-    salad: require('./gallery/salad.svg')
+    bacon: require('./svg/bacon.svg'),
+    cheese: require('./svg/cheese.svg'),
+    meat: require('./svg/meat.svg'),
+    salad: require('./svg/salad.svg')
   },
   utility: {
-    arrow: require('./gallery/arrow.svg'),
-    minus: require('./gallery/minus.svg'),
-    plus: require('./gallery/plus.svg')
+    arrow: require('./svg/arrow.svg'),
+    minus: require('./svg/minus.svg'),
+    plus: require('./svg/plus.svg')
   },
-  monster_burger_logo: require('./gallery/monster-burger.svg'),
-  crying: require('./gallery/crying.svg')
+  monster_burger_logo: require('./svg/monster-burger.svg'),
+  crying: require('./svg/crying.svg')
 }
