+++ src/components/BuildControls/BuildControl/BuildControl.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import classes from './BuildControl.module.css'
-import { images } from '../../../../assets/images'
+import { images } from '../../../assets/images'
 
 const BuildControl = ({ label, counter, addIngredient, removeIngredient, disabled, icon}) => (
   <div className={classes.BuildControl}>
