+++ src/components/Modal/Modal.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import classes from './Modal.module.css'
-import Aux from '../../hoc/Aux'
+import Aux from '../../components/Aux/Aux'
 import Backdrop from '../Backdrop/Backdrop'
 
 const Modal = ({ children, show, closeModal }) => (
