+++ src/components/ModalSummary/ModalSummary.module.css
@@ -1,15 +1,15 @@
-.OrderSummary h3 {
+.ModalSummary h3 {
   margin: 4px;
   font-weight: 600;
   font-size: 16px;
 }
 
-.OrderSummary p {
+.ModalSummary p {
   margin: 4px;
   font-size: 14px;
 }
 
-.OrderSummary ul {
+.ModalSummary ul {
   list-style: none;
   padding: 0;
   margin: 0;
@@ -17,16 +17,16 @@
   font-size: 14px;
 }
 
-.OrderSummary li {
+.ModalSummary li {
   margin: 4px;
 }
 
 @media (min-width: 550px) {
-  .OrderSummary p,
-  .OrderSummary ul {
+  .ModalSummary p,
+  .ModalSummary ul {
     font-size: 16px;
   }
-  .OrderSummary h3 {
+  .ModalSummary h3 {
     font-size: 18px;
   }
 }
