+++ src/containers/Checkout/Checkout.js
@@ -1,5 +1,5 @@
 import React, { useState, useEffect } from 'react'
-import CheckoutSummary from '../../components/Order/CheckoutSummary/CheckoutSummary'
+import CheckoutSummary from '../../components/CheckoutSummary/CheckoutSummary'
 import { Route } from 'react-router-dom'
 import ContactData from './ContactData/ContactData'
